Table: fix grouping table head bug
